updated user specific routes to check for session data

+++ flask_app/flask06.py
@@ -38,10 +38,10 @@ notes = {1:{'title': 'First note', 'text': 'This is my first note', 'date': '10-
 
 @app.route('/index')
 def index():
-    #get user from database
-    a_user =  db.session.query(User).filter_by(email='ljoffrer@uncc.edu').one()
-
-    return render_template('index.html' , user = a_user)
+    #check if a user is saved in session
+    if session.get('user'):
+        return render_template("index.html", user=session['user'])
+    return render_template("index.html")
 #^^^^^^Here we added a variable (a_user) to store our mock user data and we passed that
 # variable to our template view (index.html) with a label called user.
 
@@ -67,62 +67,66 @@ def get_note(note_id):
 #Let's add a route and function to handle requests to display this view.
 @app.route('/notes/new', methods=['GET','POST'])
 def new_note():
-    #adding commits after directory error
+    if session.get('user'):
+        if request.method == 'POST':
 
-    #check method used for request
-    if request.method == 'POST':
-        #get title data
-        title = request.form['title']
-        #get note data
-        text = request.form['noteText']
-        #create date stamp
-        from datetime import date
-        today = date.today()
-        #format date mm/dd/yyyy
-        today = today.strftime("%m-%d-%Y")
-        newEntry = Note(title, text, today)
-        db.session.add(newEntry)
-        db.session.commit()
+            title = request.form['title']
+            #get note data
+            text = request.form['noteText']
+            #create date stamp
+            from datetime import date
+            today = date.today()
+            #format date mm/dd/yyyy
+            today = today.strftime("%m-%d-%Y")
+            newEntry = Note(title, text, today, session['user_id'])
+            db.session.add(newEntry)
+            db.session.commit()
 
-        return redirect(url_for('get_notes'))
+            return redirect(url_for('get_notes'))
+        else:
+            return render_template('new.html', user=session['user'])
     else:
-        a_user = db.session.query(User).filter_by(email='ljoffrer@uncc.edu').one()
-        return render_template('new.html', user=a_user)
+        return redirect(url_for('login'))
+
+
 
 @app.route('/notes/edit/<note_id>', methods=['GET', 'POST'])
 def update_note(note_id):
     #check method used for request
-    if request.method == 'POST':
+    if session.get('user'):
+        if request.method == 'POST':
         #get title data
-        title = request.form['title']
+            title = request.form['title']
         #get note data
-        text = request.form['noteText']
-        note = db.session.query(Note).filter_by(id=note_id).one()
+            text = request.form['noteText']
+            note = db.session.query(Note).filter_by(id=note_id).one()
         #update note data
-        note.title=title
-        note.text = text
+            note.title=title
+            note.text = text
         #update note in DB
-        db.session.add(note)
+            db.session.add(note)
+            db.session.commit()
+
+            return redirect(url_for('get_notes'))
+        else:
+        #retrive note from data base:
+            my_note = db.session.query(Note).filter_by(id=note_id).one()
+
+            return render_template('new.html', note=my_note, user=session['user'])
+    else:
+        return redirect(url_for('login'))
+
+@app.route('/notes/delete/<note_id>', methods=['POST'])
+def delete_note(note_id):
+    if session.get('user'):
+        # retrieve note from database
+        my_note = db.session.query(Note).filter_by(id=note_id).one()
+        db.session.delete(my_note)
         db.session.commit()
 
         return redirect(url_for('get_notes'))
     else:
-        #get request - show new note form to edit note
-        #retrive user from data base
-        a_user= db.session.query(User).filter_by(email='ljoffrer@uncc.edu').one()
-        #retrive note from data base:
-        my_note = db.session.query(Note).filter_by(id=note_id).one()
-
-        return render_template('new.html', note=my_note, user=a_user)
-
-@app.route('/notes/delete/<note_id>', methods=['POST'])
-def delete_note(note_id):
-    # retrieve note from database
-    my_note = db.session.query(Note).filter_by(id=note_id).one()
-    db.session.delete(my_note)
-    db.session.commit()
-
-    return redirect(url_for('get_notes'))
+        return redirect(url_for('login'))
 
 #flask 6:
 @app.route('/register', methods=['POST', 'GET'])
